aula5/requirements.txt
mysqlclient==1.4.2.post1
psycopg2-binary==2.8.3
